--- Hidamari_ios/App/DailyStory.swift
@@ -1,13 +1,0 @@
-import SwiftUI
-
-@main
-struct DailyStoryApp: App {
-    @StateObject var viewModel = AuthViewModel()
-    
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-                .environmentObject(viewModel)
-        }
-    }
-}
