+++ Hidamari_ios/App/Hidamari.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+@main
+struct HidamariApp: App {
+    @StateObject var authViewModel = AuthViewModel()
+    
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+                .environmentObject(authViewModel)
+        }
+    }
+}
