+++ Hidamari_ios/Views/Components/LaunchScreenView.swift
